requirements.txt
google-adk==1.19.0
a2a-sdk==0.3.19
dotenv==0.9.9
litellm==1.80.7
psycopg2-binary==2.9.11
asyncpg==0.31.0
greenlet==3.2.4
fastmcp